Fix Index Codec memory and storage usage
